invert Y checkbox

+++ etc/scene-viewer.py
@@ -26,7 +26,7 @@ except:
 import math
 from mathutils import Matrix
 from functools import cmp_to_key
-    
+
 bl_info = {"name": "GamePlay 3D Scene Viewer", "category": "User"}
 
 bpy.types.Scene.rotatex = bpy.props.BoolProperty(name="Rotate X-90",default=False)
@@ -145,6 +145,13 @@ class SceneView(bpy.types.Operator):
             return True
     
     def execute(self, context):
+        
+        # set mode to 'OBJECT'
+        for obj in bpy.context.scene.objects:
+          if obj.type == 'MESH':
+            bpy.context.scene.objects.active = obj
+            bpy.ops.object.mode_set(mode='OBJECT')
+        
         if bpy.context.scene.rotatex:
            # Rotate -90 around the X-axis
            rotateScene(-math.pi / 2.0)
@@ -153,7 +160,7 @@ class SceneView(bpy.types.Operator):
         else:
            axisForward='Z'
            axisUp='Y'
-	
+    
         sve = bpy.context.scene.viewer_path
         svp = bpy.context.scene.game_path
         resdir = ''
@@ -207,7 +214,7 @@ class SceneView(bpy.types.Operator):
         args.append(sfp+".fbx")
         subprocess.call(args)
         for img in bpy.data.images.keys():    
-            if bpy.data.images[img].source=='FILE' and os.path.dirname(bpy.data.images[img].filepath) != resdir:
+            if bpy.data.images[img].source=='FILE' and os.path.dirname(bpy.data.images[img].filepath) != resdir and os.path.exists(bpy.data.images[img].filepath):
                 shutil.copy(bpy.data.images[img].filepath,resdir)
 
         if macApp:
@@ -216,9 +223,12 @@ class SceneView(bpy.types.Operator):
             subprocess.Popen([sve,barename],cwd=svp)
 
         if bpy.context.scene.rotatex:
-           # Rotate 90 around the X-axis
+           # Rotate 90 around the X-axis 
            rotateScene(math.pi / 2.0)
-			
+       
+           # or use undo, so it rotates scene back (and get instances back)
+           #bpy.ops.ed.undo()
+            
         return {"FINISHED"}
 
 class GameplayPanel(bpy.types.Panel):
